feat: add coupon segment registry and admin UI
- Add billing_segments model, migration, repository, and service APIs
- Validate coupon target_segment against active billing segments
- Expose admin segment endpoints for create, list, and update
- Add Coupons UI segment dropdown, segment creation dialog, and segment table
- Document coupons, promotion codes, segments, and invoice discount behavior
- Split long Casbin policy regexes to fit casbin_rule column limits

+++ db/migrations/20260317140000_rating_aggregates.up.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS rating_results (
     meter_id UUID NOT NULL REFERENCES meters(id) ON DELETE RESTRICT,
     currency TEXT NOT NULL,
     quantity NUMERIC NOT NULL CHECK (quantity >= 0),
-    unit_amount_cents BIGINT NOT NULL CHECK (unit_amount_cents >= 0),
+    unit_amount_cents NUMERIC(28,12) NOT NULL CHECK (unit_amount_cents >= 0),
     amount_cents BIGINT NOT NULL CHECK (amount_cents >= 0),
     source TEXT NOT NULL DEFAULT 'usage',
     window_start TIMESTAMPTZ NOT NULL,
